+++ Pages/Settings.xaml
@@ -1,5 +1,4 @@
-﻿<UserControl x:Class="FirstFloor.ModernUI.App.Pages.Settings"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+﻿<UserControl xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
@@ -7,11 +6,10 @@
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300">
     <Grid Style="{StaticResource ContentRoot}">
-        <mui:ModernTab SelectedSource="Pages/SettingsAppearance.xaml" Layout="List">
+        <mui:ModernTab SelectedSource="Pages/AppearanceSettings.xaml" Layout="List">
             <mui:ModernTab.Links>
-                <mui:Link DisplayName="Appearance" Source="Pages/SettingsAppearance.xaml" />
-                <mui:Link DisplayName="ScreenShot" Source="Pages/SettingsScreenShot.xaml" />
-                <mui:Link DisplayName="Video" Source="Pages/SettingsVideo.xaml" />
+                <mui:Link DisplayName="Appearance" Source="Pages/AppearanceSettings.xaml" />
+                <mui:Link DisplayName="ScreenShot" Source="Pages/ScreenShotSettings.xaml" />
             </mui:ModernTab.Links>
         </mui:ModernTab>
     </Grid>
