--- Pages/SettingsVideo.xaml
@@ -1,93 +1,0 @@
-﻿<UserControl x:Class="Captura.SettingsVideo"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:mui="http://firstfloorsoftware.com/ModernUI"
-             mc:Ignorable="d"
-             d:DesignHeight="300" d:DesignWidth="300">
-    <Grid>
-        <ScrollViewer>
-            <StackPanel x:Name="Panel" MinWidth="150">
-                <TextBlock Text="VIDEO" Style="{StaticResource Heading2}" Margin="0,0,0,8" />
-
-                <!--<Fluent:RibbonGroupBox>
-                    <Fluent:Spinner Header="Frame Rate" 
-                                    IsEnabled="{Binding ReadyToRecord}"
-                                    Margin="0,5"
-                                    x:Name="FrameRate" 
-                                    Value="10" 
-                                    Minimum="1" 
-                                    Maximum="30"/>
-                    <Fluent:Spinner Minimum="1"
-                                    Maximum="100"
-                                    IsEnabled="{Binding ReadyToRecord}"
-                                    Header="Quality"
-                                    Value="70"
-                                    Name="Quality"/>
-                </Fluent:RibbonGroupBox>
-                <Fluent:RibbonGroupBox>
-                    <Fluent:ToggleButton LargeIcon="Images/Frame.png"
-                                         Header="Region Selector"
-                                         Icon="Images/Frame.png"
-                                         Name="RegSelBox" 
-                                         Checked="ShowRegionSelector"
-                                         Unchecked="HideRegionSelector"/>
-                    <Fluent:DropDownButton Name="BgColor"
-                                           Header="Background Color"
-                                           MaxDropDownHeight="500"
-                                           Icon="{Binding LargeIcon}">
-                        <Fluent:DropDownButton.LargeIcon>
-                            <Grid>
-                                <Rectangle Height="20" 
-                                           Width="20"
-                                           Stroke="Black">
-                                    <Rectangle.Fill>
-                                        <SolidColorBrush Color="{Binding ThemeColor}" />
-                                    </Rectangle.Fill>
-                                </Rectangle>
-                            </Grid>
-                        </Fluent:DropDownButton.LargeIcon>
-                        <Fluent:ColorGallery SelectedColor="{Binding ThemeColor, Mode=TwoWay}"
-                                             Mode="ThemeColors"
-                                             StandardColorGridRows="10"
-                                             Columns="10"
-                                             IsNoColorButtonVisible="True"
-                                             IsAutomaticColorButtonVisible="False">
-                        </Fluent:ColorGallery>
-                    </Fluent:DropDownButton>
-                </Fluent:RibbonGroupBox>
-                <Fluent:RibbonGroupBox>
-                    <Fluent:ToggleButton Header="Capture Mouse Clicks"
-                                         Name="CaptureMouseClicks"
-                                         IsEnabled="{Binding ReadyToRecord}"
-                                         LargeIcon="Images/Click.png"
-                                         Icon="Images/Click.png"/>
-                    <Fluent:ToggleButton Header="Capture Keystrokes"
-                                         Name="CaptureKeyStrokes"
-                                         IsEnabled="{Binding ReadyToRecord}"
-                                         LargeIcon="Images/Keystroke.png"
-                                         Icon="Images/Keystroke.png"/>
-                </Fluent:RibbonGroupBox>-->
-
-                <CheckBox Margin="0,5" Content="Include Cursor"
-                          IsChecked="{Binding Source={StaticResource Settings}, Path=Default.IncludeCursor}"/>
-
-                <Grid HorizontalAlignment="Left">
-                    <Grid.ColumnDefinitions>
-                        <ColumnDefinition Width="Auto" />
-                        <ColumnDefinition MinWidth="120"/>
-                    </Grid.ColumnDefinitions>
-                    <Grid.RowDefinitions>
-                        <RowDefinition />
-                    </Grid.RowDefinitions>
-
-                    <TextBlock Text="Source:" VerticalAlignment="Center" Margin="0,0,16,4" />
-                    <ComboBox Name="WindowBox" Grid.Column="1" VerticalAlignment="Center" Margin="0,0,0,4"
-                              ItemsSource="{Binding AvailableWindows}" SelectedValue="{Binding _SelectedWindow, Mode=TwoWay}"
-                              SelectedValuePath="Key" DisplayMemberPath="Value"/>
-                </Grid>
-            </StackPanel>
-        </ScrollViewer>
-    </Grid>
-</UserControl>
